+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v1.2.0] - 2020-12-08
 ### Added
 - Added telnettest Server for mocking TELNET connections.
+
+### Changed
+- Replaced testify/assert to native tests.
 
 ## [v1.1.0] - 2020-11-16
 ### Added
@@ -30,6 +35,7 @@ job was joined with tests workflow.
 ### Added
 - Initial implementation.
 
-[Unreleased]: https://github.com/gorcon/telnet/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/gorcon/telnet/compare/v1.2.0...HEAD
+[v1.2.0]: https://github.com/gorcon/telnet/compare/v1.1.0...v1.2.0
 [v1.1.0]: https://github.com/gorcon/telnet/compare/v1.0.1...v1.1.0
 [v1.0.1]: https://github.com/gorcon/telnet/compare/v1.0.0...v1.0.1